Create test_repl.py and move run_pty() to test.support
