+++ Lib/test/test_readline.py
@@ -1,17 +1,11 @@
 """
 Very minimal unittests for parts of the readline module.
 """
-from contextlib import ExitStack
-from errno import EIO
 import os
-import selectors
-import subprocess
-import sys
 import tempfile
 import unittest
-from textwrap import dedent
 from test.support import (import_module, unlink, temp_dir, TESTFN, verbose,
-                          SuppressCrashReport)
+                         run_pty)
 from test.support.script_helper import assert_python_ok
 
 # Skip tests if there is no readline module
@@ -271,81 +265,6 @@ readline.write_history_file(history_file)
                 lines = f.readlines()
             self.assertEqual(len(lines), history_size)
             self.assertEqual(lines[-1].strip(), b"last input")
-
-
-# Tests for the interactive interpreter.
-class TestInteractiveInterpreter(unittest.TestCase):
-
-    def test_interactive_no_memory(self):
-        # Issue #30696: Fix the interactive interpreter looping endlessly when
-        # no memory. Check also that the fix does not break the interactive
-        # loop when an exception is raised.
-        user_input = """
-            import sys, _testcapi
-            1/0
-            _testcapi.set_nomemory(0)
-            sys.exit(0)
-        """
-        user_input = dedent(user_input)
-        user_input = b'\r'.join(x.encode() for x in user_input.split('\n'))
-        user_input += b'\r'
-        with SuppressCrashReport():
-            output = run_pty(None, input=user_input)
-        self.assertIn(b"Fatal Python error: Cannot recover from MemoryErrors",
-                      output)
-
-
-def run_pty(script, input=b"dummy input\r", env=None):
-    pty = import_module('pty')
-    output = bytearray()
-    [master, slave] = pty.openpty()
-    if script is None:
-        args = (sys.executable, '-q')
-    else:
-        args = (sys.executable, '-c', script)
-    proc = subprocess.Popen(args, stdin=slave, stdout=slave, stderr=slave, env=env)
-    os.close(slave)
-    with ExitStack() as cleanup:
-        cleanup.enter_context(proc)
-        def terminate(proc):
-            try:
-                proc.terminate()
-            except ProcessLookupError:
-                # Workaround for Open/Net BSD bug (Issue 16762)
-                pass
-        cleanup.callback(terminate, proc)
-        cleanup.callback(os.close, master)
-        # Avoid using DefaultSelector and PollSelector. Kqueue() does not
-        # work with pseudo-terminals on OS X < 10.9 (Issue 20365) and Open
-        # BSD (Issue 20667). Poll() does not work with OS X 10.6 or 10.4
-        # either (Issue 20472). Hopefully the file descriptor is low enough
-        # to use with select().
-        sel = cleanup.enter_context(selectors.SelectSelector())
-        sel.register(master, selectors.EVENT_READ | selectors.EVENT_WRITE)
-        os.set_blocking(master, False)
-        while True:
-            for [_, events] in sel.select():
-                if events & selectors.EVENT_READ:
-                    try:
-                        chunk = os.read(master, 0x10000)
-                    except OSError as err:
-                        # Linux raises EIO when slave is closed (Issue 5380)
-                        if err.errno != EIO:
-                            raise
-                        chunk = b""
-                    if not chunk:
-                        return output
-                    output.extend(chunk)
-                if events & selectors.EVENT_WRITE:
-                    try:
-                        input = input[os.write(master, input):]
-                    except OSError as err:
-                        # Apparently EIO means the slave was closed
-                        if err.errno != EIO:
-                            raise
-                        input = b""  # Stop writing
-                    if not input:
-                        sel.modify(master, selectors.EVENT_READ)
 
 
 if __name__ == "__main__":
